Entry: Access trust using a trust provider, passed from the caller.  Use this object to retrieve users to encode the entry for. This makes it so we use right user when encoding the entry.  When the client API is being used, it fallbacks to admin's trust provider.
This change fixes the test, in which an entry is being encoded for
globally-untrusted user (admin doesn't like them), but since we trust
ourselves, the entry is being encoded for us -- and only us.  Oh, yeah,
in that test we also don't trust our admin.  Stupid situation btw...

Adjust that test a bit to assert 'encoded for users' collection of the
entry.  May help probably.

git-svn-id: file:///home/repos/rinrab/MKW/trunk@4334 0083e6b4-6090-234a-836e-88fcc467669c